+++ src/TestSuite/ConsoleIntegrationTestTrait.php
@@ -30,7 +30,7 @@ trait ConsoleIntegrationTestTrait
     /**
      * @var \MeTools\Console\Command
      */
-    protected $Command;
+    protected Command $Command;
 
     /**
      * Called before every test method
@@ -40,17 +40,13 @@ trait ConsoleIntegrationTestTrait
     {
         parent::setUp();
 
-        if (preg_match('/TraitTest$/', get_class($this))) {
-            return;
-        }
-
-        if (!$this->Command && !empty($this->autoInitializeClass)) {
+        if (empty($this->Command) && !empty($this->autoInitializeClass)) {
             /** @var class-string<\MeTools\Console\Command> $className */
             $className = $this->getOriginClassNameOrFail($this);
             $this->Command = new $className();
         }
 
-        if ($this->Command && method_exists($this->Command, 'initialize')) {
+        if (!empty($this->Command) && method_exists($this->Command, 'initialize')) {
             $this->Command->initialize();
         }