+++ tests/TestCase/Command/Install/CreateDirectoriesCommandTest.php
@@ -40,9 +40,7 @@ class CreateDirectoriesCommandTest extends TestCase
 
         $dirs = Configure::read('WRITABLE_DIRS');
         $method = $Command->expects($this->exactly(count($dirs)))->method('createDir');
-        $consecutiveCalls = array_map(function (string $path) use ($io) {
-            return [$io, $path];
-        }, $dirs);
+        $consecutiveCalls = array_map(fn(string $path) => [$io, $path], $dirs);
         call_user_func_array([$method, 'withConsecutive'], $consecutiveCalls);
 
         $this->assertNull($Command->run([], $io));